JAMES-1854 Get a ride of one level of inheritance inside SieveMailet
